min_perf_check: modify to have fork in thread
`execv` replaces the whole process when executed. A fork is therefore
mandatory. It has been put in a thread in order to lighten the load of
the fork but also to restrict its area.
Our thread now returns the time as well as it is being calculated during
the thread.